Push revision to kafka topic and minor cleanup
