+++ apply-rule/src/main/resources/application.json
@@ -3,6 +3,6 @@
     "localhost:9092"
   ],
   "receiver_topic": "xadf.compute.items",
-  "producer_topic": "xadf.result.items",
+  "producer_topic": "xadf.result.documents",
   "cassandra_host": "localhost"
 }
